bl_basename: Allow const string argument
There are users ob basename that pass in a const string.

This is an error with gcc14. It's possible to make a copy of the string
before passing it to basename but there is also another way: avoid
modifiyng the string in basename.

basename can return pointer to the start of the base part without
changing the string. However, it cannot strip trailing slashes.