Fix bulk allocation warnings
mlterm uses a trick that the C standrds committee hates: it allocates a
structure and its members as one memory block so it can be freed as one
block without complex recursive freeing of members.

To vaoid an error with gcc14 instead of assigning pointer to the
structure increased by one to its member assign the malloc result into a
temporary void pointer, and use void pointer arithmetic to get the
allocated member address.

Another variant is allocating a memory block and assigning it to
multiple incompatible pointes at once.